SHBShop/shbshop_back/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.0.2
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
more-itertools==10.6.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.0
python-dotenv==1.1.0
sqlacodegen==3.0.0
SQLAlchemy==2.0.40
typeguard==4.4.2
typing_extensions==4.13.0
Werkzeug==3.1.3
